pipeline/pyproject.toml
[project]
name = "pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.2",
    "asyncpg>=0.31.0",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0",
    "tenacity>=9.1.2",
]
